20.12.2016

+++ admin/model/tool/mysklad_oc23.php
@@ -114,65 +114,6 @@ class ModelToolMysklad_oc23 extends Model {
 					$product_counter++;
 				}
 				
-				//Доставка
-							$totals = $this->model_sale_order->getOrderTotals($orders_data['order_id']);
-
-							foreach ($totals as $total) {
-						 if ($total['code']=='shipping') {
-
-							$document['Документ' . $document_counter]['Товары']['Товар' . $product_counter] = array(
-									 'Ид'         => ''
-									,'Наименование' => 'Доставка'
-									,'ЦенаЗаЕдиницу'=> $total['value']
-									,'Количество' => 1
-									,'Сумма'       => $total['value']
-								);
-							}
-
-							}
-				$data = $order;
-				
-				//Статус
-					 
-					$order_status = $this->model_sale_order->getOrderStatus($orders_data['order_id']); 
-								
-					$document['Документ' . $document_counter]['ЗначенияРеквизитов']['ЗначениеРеквизита0']= array(
-							  
-								'Наименование' => 'Статус заказа',
-								'Значение'=> $order_status,
-								
-							
-						);
-						
-				//Метод оплаты
-					 
-					$pay_status = $this->model_sale_order->getOrderPay($orders_data['order_id']); 
-								
-					$document['Документ' . $document_counter]['ЗначенияРеквизитов']['ЗначениеРеквизита1'] = array(
-							
-								'Наименование' => 'Метод оплаты',
-								'Значение'=> $pay_status,
-							
-						);
-						
-				//Способ доставки
-					 
-					$shipping_status = $this->model_sale_order->getOrderShipping($orders_data['order_id']); 
-								
-					$document['Документ' . $document_counter]['ЗначенияРеквизитов']['ЗначениеРеквизита2']= array(
-							
-								'Наименование' => 'Способ доставки',
-								'Значение'=> $shipping_status,
-							
-						);
-						
-
-				$this->model_sale_order->addOrderHistory($orders_data['order_id'], array(
-					'order_status_id' => $params['new_status'],
-					'comment'         => '',
-					'notify'          => $params['notify']
-				));
-
 				$document_counter++;
 			}
 			
@@ -182,6 +123,28 @@ class ModelToolMysklad_oc23 extends Model {
 		$xml = $this->array_to_xml($document, new SimpleXMLElement($root));
 
 		return $xml->asXML();
+	}
+public function queryOrdersStatus($params){
+
+		$this->load->model('sale/order');
+
+		if ($params['exchange_status'] != 0) {
+			$query = $this->db->query("SELECT order_id FROM `" . DB_PREFIX . "order` WHERE `order_status_id` = " . $params['exchange_status'] . "");
+		} else {
+			$query = $this->db->query("SELECT order_id FROM `" . DB_PREFIX . "order` WHERE `date_added` >= '" . $params['from_date'] . "'");
+		}
+
+		if ($query->num_rows) {
+			foreach ($query->rows as $orders_data) {
+				$this->model_sale_order->addOrderHistory($orders_data['order_id'], array(
+					'order_status_id' => $params['new_status'],
+					'comment'         => '',
+					'notify'          => $params['notify']
+				));
+			}
+		}
+
+		return true;
 	}
 
 
@@ -221,7 +184,7 @@ class ModelToolMysklad_oc23 extends Model {
 		$xml = simplexml_load_file($importFile);
 		
 		$price_types = array();
-		
+		$config_price_type_main = array();
 		$enable_log = $this->config->get('mysklad_oc23_full_log');
 		$mysklad_oc23_relatedoptions = $this->config->get('mysklad_oc23_relatedoptions');
 
@@ -230,14 +193,17 @@ class ModelToolMysklad_oc23 extends Model {
 		if ($enable_log)
 			$this->log->write("Начат разбор файла: " . $filename);
 
-		if ($xml->ПакетПредложений->ТипыЦен->ТипЦены) {
-			foreach ($xml->ПакетПредложений->ТипыЦен->ТипЦены as $type) {
-				$price_types[(string)$type->Ид] = (string)$type->Наименование;
-			}
-		}
-
 		if (!empty($config_price_type) && count($config_price_type) > 0) {
 			$config_price_type_main = array_shift($config_price_type);
+		}
+
+		if ($xml->ПакетПредложений->ТипыЦен->ТипЦены) {
+			foreach ($xml->ПакетПредложений->ТипыЦен->ТипЦены as $key => $type) {
+				$price_types[(string)$type->Ид] = (string)$type->Наименование;
+				if($key == 0 && count($config_price_type_main) == 0) {
+					$config_price_type_main['keyword'] = (string)$type->Наименование;
+				}
+			}
 		}
 
 		// Инициализация массива скидок для оптимизации алгоритма
@@ -420,7 +386,12 @@ class ModelToolMysklad_oc23 extends Model {
 						}
 					}
 	
-					$data['status'] = 1;
+					if ($offer->Статус) {
+						$data['status'] = (string)$offer->Статус;
+					}
+					else {
+						$data['status'] = 1;
+					}
 				}
 				
 				if (!$mysklad_oc23_relatedoptions || $offer_cnt == count($xml->ПакетПредложений->Предложения->Предложение)
@@ -506,7 +477,7 @@ class ModelToolMysklad_oc23 extends Model {
 
 				$data['model'] = $product->Артикул? (string)$product->Артикул : 'не задана';
 				$data['name'] = $product->Наименование? (string)$product->Наименование : 'не задано';
-				$data['weight'] = $product->Вес? (float)$product->Вес : 0;
+				$data['weight'] = $product->Вес? (float)$product->Вес : null;
 				$data['sku'] = $product->Артикул? (string)$product->Артикул : '';
 
 				if ($enable_log)
@@ -526,6 +497,7 @@ class ModelToolMysklad_oc23 extends Model {
 				if($product->ХарактеристикиТовара){
 
 					$count_options = count($product->ХарактеристикиТовара->ХарактеристикаТовара);
+					$option_desc = '';
 
 					foreach($product->ХарактеристикиТовара->ХарактеристикаТовара as $option ) {
 						$option_desc .= (string)$option->Наименование . ': ' . (string)$option->Значение . ';';
@@ -533,9 +505,14 @@ class ModelToolMysklad_oc23 extends Model {
 					$option_desc .= ";\n";
 				}
 
-				if ($product->Группы) $data['category_1c_id'] = (string)$product->Группы->Ид;
+				if ($product->Группы) $data['category_1c_id'] = $product->Группы->Ид;
 				if ($product->Описание) $data['description'] = (string)$product->Описание;
-				if ($product->Статус) $data['status'] = (string)$product->Статус;
+				if ($product->Статус) {
+					$data['status'] = (string)$product->Статус;
+				}
+				else {
+					$data['status'] = 1;
+				}
 
 				// Свойства продукта
 				if ($product->ЗначенияСвойств) {
@@ -585,6 +562,11 @@ class ModelToolMysklad_oc23 extends Model {
 
 										$manufacturer_id = $this->model_catalog_manufacturer->addManufacturer($data_manufacturer);
 										$data['manufacturer_id'] = $manufacturer_id;
+										//только если тип 'translit'
+										if ($this->config->get('exchange1c_seo_url') == 2) {
+											$man_name = "brand-" . $manufacturer_name;
+											$this->setSeoURL('manufacturer_id', $manufacturer_id, $man_name);
+										}
 									}
 								break;
 
@@ -671,7 +653,8 @@ class ModelToolMysklad_oc23 extends Model {
 				,'meta_keyword'     => (isset($data['category_description'][$language_id]['meta_keyword'])) ? $data['category_description'][$language_id]['meta_keyword'] : ''
 				,'meta_description'	=> (isset($data['category_description'][$language_id]['meta_description'])) ? $data['category_description'][$language_id]['meta_description'] : ''
 				,'description'		  => (isset($category->Описание)) ? (string)$category->Описание : ((isset($data['category_description'][$language_id]['description'])) ? $data['category_description'][$language_id]['description'] : '')
-				,'seo_title'        => (isset($data['category_description'][$language_id]['seo_title'])) ? $data['category_description'][$language_id]['seo_title'] : ''
+				,
+				'meta_title'        => (isset($data['category_description'][$language_id]['seo_title'])) ? $data['category_description'][$language_id]['seo_title'] : ''
 				,'seo_h1'           => (isset($data['category_description'][$language_id]['seo_h1'])) ? $data['category_description'][$language_id]['seo_h1'] : ''
 			)
 		);
@@ -716,6 +699,12 @@ class ModelToolMysklad_oc23 extends Model {
 				$this->CATEGORIES[$id] = $category_id;
 			}
 
+			//только если тип 'translit'
+			if ($this->config->get('exchange1c_seo_url') == 2) {
+				$cat_name = "category-" . $data['parent_id'] . "-" . $data['category_description'][$language_id]['name'];
+				$this->setSeoURL('category_id', $category_id, $cat_name);
+			}
+
 			if ($category->Группы) $this->insertCategory($category->Группы->Группа, $category_id, $language_id);
 		}
 
@@ -731,12 +720,13 @@ class ModelToolMysklad_oc23 extends Model {
 	private function insertAttribute($xml) {
 		$this->load->model('catalog/attribute');
 		$this->load->model('catalog/attribute_group');
+		$lang_id = (int)$this->config->get('config_language_id');
 
 		$attribute_group = $this->model_catalog_attribute_group->getAttributeGroup(1);
 
 		if (!$attribute_group) {
 
-			$attribute_group_description[1] = array (
+			$attribute_group_description[$lang_id] = array (
 				'name' => 'Свойства'
 			);
 
@@ -755,11 +745,12 @@ class ModelToolMysklad_oc23 extends Model {
 			//Дима, тест
 			$optionValue = $attribute->ТипыЗначений->ТипЗначений;
 			//Дима, тест!
-			if ((string)$optionValue->ВариантыЗначений) {
-				if ((string)$optionValue->Тип/*ТипЗначений*/ == 'Справочник') {
-					foreach($optionValue->ВариантыЗначений/*->Справочник*/ as $option_value){
-						if ((string)$option_value->ВариантЗначения->Значение != '') {
-							$values[(string)$option_value->ВариантЗначения->ИдЗначения] = (string)$option_value->ВариантЗначения->Значение;
+			
+			if ((string)$attribute->ВариантыЗначений) {
+				if ((string)$attribute->ТипЗначений == 'Справочник') {
+					foreach($attribute->ВариантыЗначений->Справочник as $option_value){
+						if ((string)$option_value->Значение != '') {
+							$values[(string)$option_value->ИдЗначения] = (string)$option_value->Значение;
 						}
 					}
 				}
@@ -770,7 +761,7 @@ class ModelToolMysklad_oc23 extends Model {
 				'sort_order'            => 0,
 			);
 
-			$data['attribute_description'][1]['name'] = (string)$name;
+			$data['attribute_description'][$lang_id]['name'] = (string)$name;
 
 			// Если атрибут уже был добавлен, то возвращаем старый id, если атрибута нет, то создаем его и возвращаем его id
 			$current_attribute = $this->db->query('SELECT attribute_id FROM ' . DB_PREFIX . 'attribute_to_1c WHERE 1c_attribute_id = "' . $id . '"');
@@ -935,13 +926,28 @@ class ModelToolMysklad_oc23 extends Model {
 			$product['product_option'] = array();
 		}
 
-		if (isset($product['category_1c_id']) && isset($this->CATEGORIES[$product['category_1c_id']])) {
-			$result['product_category'] = array((int)$this->CATEGORIES[$product['category_1c_id']]);
-			$result['main_category_id'] = (int)$this->CATEGORIES[$product['category_1c_id']];
+		if (isset($product['category_1c_id'])) {
+			if (is_object($product['category_1c_id'])) {
+				foreach ($product['category_1c_id'] as $category_item) {
+					if (isset($this->CATEGORIES[(string)$category_item])) {
+						$result['product_category'][] = (int)$this->CATEGORIES[(string)$category_item];
+						$result['main_category_id'] = 0;
+					}
+				}
+			} else {
+				$product['category_1c_id'] = (string)$product['category_1c_id'];
+				if (isset($this->CATEGORIES[$product['category_1c_id']])) {
+					$result['product_category'] = array((int)$this->CATEGORIES[$product['category_1c_id']]);
+					$result['main_category_id'] = (int)$this->CATEGORIES[$product['category_1c_id']];
+				} else {
+					$result['product_category'] = isset($data['product_category']) ? $data['product_category'] : array(0);
+					$result['main_category_id'] = isset($data['main_category_id']) ? $data['main_category_id'] : 0;
+				}
+			}
 		}
-		else {
-			$result['product_category'] = isset($data['product_category']) ? $data['product_category']: array(0);
-			$result['main_category_id'] = isset($data['main_category_id']) ? $data['main_category_id']: 0;
+		
+		if (!isset($result['product_category']) && isset($data['product_category'])) {
+			$result['product_category'] = $data['product_category'];
 		}
 		
 		if (isset($product['related_options_use'])) {
@@ -976,17 +982,23 @@ class ModelToolMysklad_oc23 extends Model {
 			$this->updateProduct($product, $product_id, $language_id);
 		}
 		else {
-			// Проверяем, существует ли товар с тем-же артикулом
-			// Если есть, то обновляем его
-			$product_id = $this->getProductBySKU($data['sku']);
-			if ($product_id !== false) {
-				$this->updateProduct($product, $product_id, $language_id);
-			}
-			// Если нет, то создаем новый
-			else {
+
+			if ($this->config->get('mysklad_oc23_dont_use_artsync')) {
 				$this->load->model('catalog/product');
-				$this->model_catalog_product->addProduct($data);
+				$product_id =	$this->model_catalog_product->addProduct($data);
+			} else {
+				// Проверяем, существует ли товар с тем-же артикулом
+				// Если есть, то обновляем его
 				$product_id = $this->getProductBySKU($data['sku']);
+				if ($product_id !== false) {
+					$this->updateProduct($product, $product_id, $language_id);
+				}
+				// Если нет, то создаем новый
+				else {
+					$this->load->model('catalog/product');
+					$this->model_catalog_product->addProduct($data);
+					$product_id = $this->getProductBySKU($data['sku']);
+				}
 			}
 
 			// Добавляем линк
@@ -994,6 +1006,14 @@ class ModelToolMysklad_oc23 extends Model {
 				$this->db->query('INSERT INTO `' .  DB_PREFIX . 'product_to_1c` SET product_id = ' . (int)$product_id . ', `1c_id` = "' . $this->db->escape($product['1c_id']) . '"');
 			}
 		}
+		// Устанавливаем SEO URL
+		if ($product_id){
+			//только если тип 'translit'
+			if ($this->config->get('exchange1c_seo_url') == 2) {
+				$this->setSeoURL('product_id', $product_id, $product['name']);
+			}
+		}
+
 	}
 
 
@@ -1045,6 +1065,37 @@ class ModelToolMysklad_oc23 extends Model {
 		//Редактируем продукт
 		$product_id = $this->model_catalog_product->editProduct($product_id, $product_old);
 
+	}
+
+	/**
+	 * Устанавливает SEO URL (ЧПУ) для заданного товара
+	 *
+	 * @param 	inf
+	 * @param 	string
+	 */
+	private function setSeoURL($url_type, $element_id, $element_name) {
+		$this->db->query("DELETE FROM `" . DB_PREFIX . "url_alias` WHERE `query` = '" . $url_type . "=" . $element_id . "'");
+		$this->db->query("INSERT INTO `" . DB_PREFIX . "url_alias` SET `query` = '" . $url_type . "=" . $element_id ."', `keyword`='" . $this->transString($element_name) . "'");
+	}
+
+	/**
+	 * Транслиетрирует RUS->ENG
+	 * @param string $aString
+	 * @return string type
+	 */
+	private function transString($aString) {
+		$rus = array(" ", "/", "*", "-", "+", "`", "!", "@", "#", "$", "%", "^", "&", "*", "(", ")", "_", "+", "[", "]", "{", "}", "~", ";", ":", "'", "\"", "<", ">", ",", ".", "?", "А", "Б", "В", "Г", "Д", "Е", "З", "И", "Й", "К", "Л", "М", "Н", "О", "П", "Р", "С", "Т", "У", "Ф", "Х", "Ъ", "Ы", "Ь", "Э", "а", "б", "в", "г", "д", "е", "з", "и", "й", "к", "л", "м", "н", "о", "п", "р", "с", "т", "у", "ф", "х", "ъ", "ы", "ь", "э", "ё",  "ж",  "ц",  "ч",  "ш",  "щ",   "ю",  "я",  "Ё",  "Ж",  "Ц",  "Ч",  "Ш",  "Щ",   "Ю",  "Я");
+		$lat = array("-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-", "-",  "-", "-", "-", "-", "-", "-", "a", "b", "v", "g", "d", "e", "z", "i", "y", "k", "l", "m", "n", "o", "p", "r", "s", "t", "u", "f", "h", "",  "i", "",  "e", "a", "b", "v", "g", "d", "e", "z", "i", "j", "k", "l", "m", "n", "o", "p", "r", "s", "t", "u", "f", "h", "",  "i", "",  "e", "yo", "zh", "ts", "ch", "sh", "sch", "yu", "ya", "yo", "zh", "ts", "ch", "sh", "sch", "yu", "ya");
+
+		$string = str_replace($rus, $lat, $aString);
+
+		while (mb_strpos($string, '--')) {
+			$string = str_replace('--', '-', $string);
+		}
+
+		$string = strtolower(trim($string, '-'));
+
+		return $string;
 	}
 
 	/**
@@ -1142,6 +1193,12 @@ class ModelToolMysklad_oc23 extends Model {
 	 * Заполняет продуктами родительские категории
 	 */
 	public function fillParentsCategories() {
+		$this->load->model('catalog/product');
+		if (!method_exists($this->model_catalog_product, 'getProductMainCategoryId')) {
+			$this->log->write("  !!!: Заполнение родительскими категориями отменено. Отсутствует main_category_id.");
+			return;
+		}
+		
 		$this->db->query('DELETE FROM `' .DB_PREFIX . 'product_to_category` WHERE `main_category` = 0');
 		$query = $this->db->query('SELECT * FROM `' . DB_PREFIX . 'product_to_category` WHERE `main_category` = 1');
